feat(types): migrate types to separate logic
