+++ packages/client-redes/src/graphql/FetchIndividuals.tsx
@@ -3,6 +3,7 @@ import { gql } from "apollo-boost";
 import { useSession, useQuery } from 'bonde-core-tools';
 import { useFilterQuery } from "./FilterQuery";
 import Empty from '../components/Empty';
+import { IndividualData, IndividualVars } from '../types/Individual'
 
 const USERS = gql`
   query RedeIndividuals(
@@ -58,43 +59,6 @@ const USERS = gql`
     updated_at
   }
 `;
-
-export type Individual = {
-  id: number;
-  first_name: string;
-  last_name: string;
-  email: string;
-  whatsapp: string;
-  phone: string;
-  zipcode: string;
-  address: string;
-  city: string;
-  coordinates: Record<string, string>;
-  state: string;
-  status: string;
-  availability: string;
-  extras: Record<string, string>;
-  form_entry_id: number;
-  group: {
-    id: number;
-    community_id: number;
-    is_volunteer: boolean;
-  };
-  created_at: string;
-  updated_at: string;
-};
-
-type IndividualVars = {
-  context: {
-    _eq: number;
-  };
-  filters: Record<string, string>;
-  is_volunteer: boolean;
-};
-
-interface IndividualData {
-  rede_individuals: Individual[];
-}
 
 const FetchIndividuals = ({ children, community }: any) => {
   const { filters, changeFilters, page } = useFilterQuery();
