+++ packages/client-redes/src/types/index.ts
@@ -1,1 +1,3 @@
 export * from "./Settings";
+export * from "./Groups";
+export * from "./Individual";
